+++ internal/command/views/output.go
@@ -84,13 +84,13 @@ func (v *OutputHuman) Output(name string, outputs map[string]*states.OutputValue
 		sort.Strings(ks)
 
 		for _, k := range ks {
-			v := outputs[k]
-			if v.Sensitive {
+			vs := outputs[k]
+			if vs.Sensitive && !v.view.showSensitive {
 				outputBuf.WriteString(fmt.Sprintf("%s = <sensitive>\n", k))
 				continue
 			}
 
-			result := repl.FormatValue(v.Value, 0)
+			result := repl.FormatValue(vs.Value, 0)
 			outputBuf.WriteString(fmt.Sprintf("%s = %s\n", k, result))
 		}
 	}